[Proxying] Use a dedicated worker to pass messages between threads
Rather than forwarding messages from one pthread to another through the main
thread, forward them through a dedicated worker, decreasing message latency in
cases where the main thread is busy. The new message broker worker is spawned
via a data URL so there are no new JS files for users to distribute. Whenever a
pthread is created, a new MessageChannel is created alongside it, with one
MessagePort sent to the broker and the other MessagePort sent to the pthread's
worker. The broker receives messages on these MessagePorts and forwards them on
to their recipients' MessagePorts.

It is possible for the broker to receive a message for a thread for which the
broker has no MessagePort. This can happen when pthread_create returns a
pthread_t before the main thread has finished asynchronously spinning up a new
worker to run the thread. The spawning thread may then immediately proxy work to
the new thread, causing a message to be sent to the broker before the main
thread has notified the broker of the new thread's existence and sent the broker
the new thread's MessagePort. When this happens the broker buffers the message
and forwards it along once it has received the recipient's MessagePort.

When a pthread exits or is cancelled, it closes its MessagePort so that no
further messages will be received on it, then the main thread notifies the
broker that the thread has been destroyed so the broker can release its
resources for that thread. Messages that make it to the broker after the exiting
thread has closed its MessagePort will be silently dropped or sent to the next
thread spawned with the same ID, but we have always considered it user error to
proxy to an exiting thread, so it's unclear whether doing anything better in
that case would be worth it.