Remove unused (and unusable) emscripten_sync_run_in_main_thread API.
This family of function was not used anywhere and is not currently
usable externally.

The core function `emscripten_sync_run_in_main_thread` was not usable
since the its first argument is `em_queued_call` which is not struct
that is declared externally.

The `emscripten_sync_run_in_main_thread_N` function takes a function
enum as arg0 which is not designed to be user facing.

It looks like these function were once used by JS library code but the
last remaining use of them was removed in #16405.

See: #15756