Remove unneeded exportedAsmFunc helper. NFC
It looks like this was originally added back in #8040, but was likely
not needed.

At least in of the of three remaining places this function is used
we don't need to go via the `Module` object and these symbols can/should
simply be referenced via the current scope.

Split out from #18540