Use `wasmImports` as the single global symbol table for dynamic linking (#18540)
Previously we were looking up symbols on the `Module` object which
meant building with `EXPORT_ALL` and exporting all symbols there.  This
change adds all needed symbols to the `wasmImports` maps and uses that
as the single source for symbol lookups.

This is split out from my work on #18376.

This should be a code size win in for most users of dynamic linking.  It
also avoids extra symbols being present on the `Module` object that the
user didn't ask for.