[EH] Add EXCEPTION_STACK_TRACES option
This adds `EXCEPTION_STACK_TRACES` option, which embeds stack traces
into exception objects even when `ASSERTIONS` is not set. This is for
the users who wants exception stack traces but don't want to incur the
full overhead of `ASSERTIONS`. Exception stack traces are enabled when
either of `ASSERTIONS` or `EXCEPTION_STACK_TRACES` is true.

This currently works only for Wasm EH. The reason this option's name is
not `WASM_EXCEPTION_STACK_TRACES` is I think this can work for
Emscripten EH later if something like #18535 lands.

Fixes #18533.