[Promise] Add a C API for promises (#18598)
Add a new header, emscripten/promise.h, providing a C API for creating,
resolving, chaining, and combining promises from C with a focus on providing the
full expressive capability of JS promises. To provide maximum flexibility to
users, promises can be fulfilled or rejected with arbitrary user-provided
`void*` pointers as the value or reason, or they can be resolved by providing
the handle of another promise whose eventual state should be matched. The
callbacks provided to `emscripten_promise_then` additionally receive a context
`data` pointer provided by the user as part of the `emscripten_promise_then`
call.

Remove the less expressive C API for creating promises we had in eventloop.h.
That API allowed promises to be created and resolved, but was more complicated
and less capable.