[EH/SjLj] Make Wasm EH force Wasm SjLj (#18692)
Currently `SUPPORT_LONGJMP` defaults to `emscripten` unless explicitly
otherwise. So when a user specified `-fwasm-exceptions`, it resulted in
Wasm EH + Emscripten SjLj, which has many restrictions (See
https://github.com/emscripten-core/emscripten/discussions/17526#discussioncomment-3265159
for details) and is the combination we do not intend to support for the
long term.

This PR makes `SUPPORT_LONGJMP` default value depend on the exception
mode. If Wasm EH (`-fwasm-exception`) is used, it defaults to `wasm`,
and if Emscripten EH (`-sDISABLE_EXCEPTION_CATCHING=0`) is used or no
exception support is used, it defaults to `emscripten`. This way users
get to use the same kind of EH + SjLj, which is easier to maintain and
has less restrictions.