Ensure exports are correctly handled when instantiateWasm is present (#18711)
Split out from #18705

I noticed while working on #18705 that the exports returned by
`instantiateWasm` (which is used in pthreads to load the wasm module)
ere not the same (potentially wrapped/modified) exports that were
created during `receiveInstance`.

I guess someone else must have noticed this and added the extra
call to `Asyncify.instrumentWasmExports`.  This should not be necessary
since `receiveInstance` already takes care of that.