shell.js: use === and !== where appropriate
Usage of triple vs. double equalities in shell.js was inconsistent. In all cases where `==` and `!=` were used in the file, `===` and `!==` are more appropriate.

At a minimum, the non-coercive equality checks show intention more clearly. They also are generally faster, though this is unlikely to matter in shell.js.