Avoid calling `process.exit` under node
Instead we always prefer to let exceptions bubble out and get handles by
whatever handlers are installed.

Doing this did exposed a few places where we were calling back into
the runtime after it had exited.  This was previously being masked by
the fact that we were bringing down the entire node process.  This means
that when running tests under node we will now be a little more
sensitive to use-after-exit bugs in our runtime and test code, but I
would argue this is good thing.

This also means we no longer need to define logExceptionOnExit under
node, which is node code size saving.