Track runtime keepalive status even when EXIT_RUNTIME=0
This is a partial revert of #16339, which tried to save codesize by not
tracking this information when `EXIT_RUNTIME=0`.  However this proved to
be problematic in the case of pthreads where each thread does need to
trace is keepalive status, even if the application overall never exits.

The alternative to this change would have been to add yet more
complexity, tracking the status when EXIT_STATUS || USE_PTHREADS, but
this code is already the source of many subtle bugs so keeping it simple
at the cost of a few bytes of code size seemed like a better solution
(at least for now).

Fixes: #18782, #18667