Update Harfbuzz port from 3.2.0 to 7.0.1.
This now pulls from the GitHub tag which is a `.tar.gz`, so
the older issue with the release being a `.xz` is no longer
an issue.

This also uses the unity build file, `harfbuzz.cc` rather
than dealing with each of the individual `.cc` files
on our own.

Fixes #18805.