[Proxying] Add proxying functions that return promises
Add `emscripten_proxy_promise` and `emscripten_proxy_promise_with_ctx` that
return `em_promise_t` handles to promises that are fulfilled when the proxied
work is completed or rejected when the target thread dies before it can complete
the work.

Do not document these new APIs yet since they depend on the promise API in
emscripten/promise.h, which is not yet documented. Similarly, do not add a C++
wrapper yet because emscripten/promise.h does not yet have a C++ API.