[Proxying][NFC] Move deferred cleanup logic to `em_task_queue`
We previously had a scheme to defer freeing of `em_proyxing_queue`s until there
were no more outstanding references to their `em_task_queues` sitting in
message queues. Since the notification messages only ever contain references to
`em_task_queue`s and not the `em_proxying_queue`s, move the deferred cleanup
logic to the `em_task_queue` layer from the `em_proxying_queue` layer. This
slightly simplifies the code and is a cleaner separation of concerns.

This is NFC as far as users are concerned, and the code is moved with only two
changes to internal behavior. First, we now use a trylock when culling zombies
to avoid blocking when multiple threads are creating `em_task_queue`s at the
same time. Second, we now enqueue zombie queues at the tail of the zombie list
instead of the head because FIFO behavior seems fairer.

The test for the zombie culling behavior is necessarily less precise now because
it the objects with deferred cleanup are no longer directly user-visible. The
test still does its job, though.