JSPI - Support async embind function bindings.
Adds a new `async()` option for function and class method bindings.
Functions marked async will use a separate async invoker. The invoker
will be exported and transformed in the JSPI binaryen pass. When the
binding is initialized on start up, a promising WebAssembly Function will
be created to wrap the export and extra code will be generated to handle
the promise.

Depends on https://github.com/WebAssembly/binaryen/pull/5519