[Proxying] Send messages via in-memory mailbox queues
Threads were previously notified of new work via postMessage messages that
carried pointers to the task queues to execute. There was no way to synchronously
pump or inspect these pending messages however, and there is no central registry
of all task queues for a thread, so this mechanism afforded no way to discover
or cancel pending work when a thread dies.

In preparation for implementing work cancellation, move the pending messages
into userspace by giving each thread a "mailbox", which is an `em_task_queue` in
the pthread struct. Instead of using `postMessage`, proxying queues now use the
thread mailbox API to notify threads of new work.

Internally, thread mailboxes still use postMessage to schedule work to be
executed when a thread returns to its event loop. Since the only task queues
involved in postMessages are now at known locations relative to the pthread
struct, there is no longer any need to store pointers to them in the postMessage
messages themselves. Removing these pointers works around tricky notification
and lifetime management edge cases that would have caused problems such as
dropped work or use-after-free bugs in future PRs.

When a thread dies because it exits or is canceled, it "closes" its mailbox by
decrementing a refcount and waiting to observe a refcount of 0. At this point,
the thread mailbox API ensures that no new messages will be enqueued on the
mailbox. Because the postMessage messages no longer contain task queue pointers,
it is safe to destroy the mailbox immediately after it is closed.

A user-visible behavior change this introduces is that proxied work is more
frequently completed _before_ a thread's main function begins running, since it
no longer gets ordered behind the `run` message in the JS postMessage queue. A
few tests are updated accordingly.