Remove IN_TEST_HARNESS setting
This was set my some unit tests but not all the have a set of disparate
effects on the generated code.

Better to avoid having the test hardness be able have these kinds of
special privileges, and it turns out they could all be implemented in
other ways, or were not actually needed.