Fix DECLARE_ASM_MODULE_EXPORTS=0 for mangled main function (#18895)
With `DECLARE_ASM_MODULE_EXPORTS` the list of wasm exports is statically
generated at build time.  This means that mangling in the compiler takes
care of converted `__main_argc_argv` to `main`.  e.g.  something like
the following is statically generated:

```
var main = Module['main] = asm['__main_argc_argv'];
```

However, with `DECLARE_ASM_MODULE_EXPORTS=0` the exports are dynamically
propagated from the wasm exports to the Module object, so we need to
apply the same demangling at runtime as we do at build time.

The build time version of this logic is here:
https://github.com/emscripten-core/emscripten/blob/c4693feae0e366cf7741e9e1725fda5539605d4d/tools/shared.py#L656-L657