[Proxying] Use Atomics.waitAsync to receive proxying notifications (#18861)
When it is available, register to be notified of new proxying messages using
`Atomics.waitAsync` and perform the notification using `Atomics.notify`. This
new notification mechanism does not depend on the main thread forwarding
postMessage notifications, so its performance should not degrade when the main
thread is busy.

Update a test that depended on intercepting postMessage notifications to
conservatively wait for 20ms instead.