Add tests that use emscripten output in a bundler (webpack). NFC
If we want to iterate and improve on or support for MODULARIZE I think
its good it we start testing it in the environments where it will
likely be used.

This change and an initial test for the use of MODULARIZE with the
webpack bundler and verifies that the resulting bundle works.