Ninja library build: allow deferred/combined builds using a parent ninja (#17871)
This creates separate build/generate steps for library builds, and uses them in embuilder
when building multiple libraries at once (non-embuilder builds are still separated).

It creates a "parent" build.ninja and adds a subninja for each library as it is built (either
by embuilder or emcc). Embuilder gets a "rebuild" command in embuilder that
invokes ninja on that file, thus rebuilding all libraries that have been built in the
current cache.

This is expected to be most useful for developing emscripten (since the dependencies
will ensure only necessary files are rebuilt when editing emscripten library sources), but
using it on CircleCI also cuts 7 minutes (~25%) off build-linux step due to parallelism
across libraries.