Automatically generate `__sig` properties for JS library symbols
This primary change here is the creation of `emscripten_internal.h`
which is an internal-only header that contains that declaration of
internal JS function that are not already declared in other files.

The reason for this is to ensure that all JS library functions that are
callable form native code have a valid declaration, at least internally,
so that gen_sig_info.py can extract that signature from a compiled wasm
file.

This change add the scripte for generating `library_sigs.js` but doesn't
actually start using it, or erase the correpsonding hand-created `__sig`
entries. That is will be part of a followup