Update SDL_image port to 2.6.0
This seems to be the latest release that exists on the SDL2 branch
of the upstream repository.

I verified that both of own downstream patches that we had in the
emscripten-ports fork have since landed upstream:

https://github.com/libsdl-org/SDL_image/commit/1d92df494
https://github.com/libsdl-org/SDL_image/commit/840782235

I also verified that both of these made it into the 2.6.0 release.

This change is needed as part of #19004 but it also allows us to
delete our fork which is nice.