Store generated struct info rather than auto-generating it in the sysroot. NFC (#19013)
I noticed while working `gen_sig_info.py` in #18985 that is was simpler
to just keep a copy of the generated file checked into source control.
We can use a unit test check that its up-to-date on each commit.  In
fact we already had a unit test that was doing this and keeping a copy
of the expected output in the test directory.

This simplifies the emcc linker code since we can always assume the
presence of the struct info file.

Also allows us to remove the `varies=False` special case in cache logic.