Avoid GL symbols being added to library as "undefined". NFC
The `recordGLProcAddressGet` was processing aliases, but it was
assuming the aliases existed within the same library file which in
was not the case for `glGenFramebuffersOES` and others in this file
that alias symbols in `library_webgl.js`.

This was resulting the jsifier thinking the symbols was defined (since
it existed in the library) but generating undefined symbols in the
in the final output such as:

```
var glGenFramebuffersOES = undefined;
```

Removing the aliasing resolution from `recordGLProcAddressGet` lets
`jsifier.js` handle the alias which results in the following output
after this change:

```
var _glGenFramebuffersOES = _glGenFramebuffers;
```

Also add some assertions so these types of error don't get overlooked in
the future.