jsifier: Add explicit error messages for missing/native alias targets
Followup to #19033.

Updating the `test_closure_full_js_library` to include webgl2 caused
a lot of `JSC_REFERENCE_BEFORE_DECLARE` closure errors.

This is because we had JS aliases who's target was defined both in
JS and in native code.  In this case the jsifier chooses not to include
the JS version but the location where the aliases are declared in the
output comes before the native symbols are declared.

To fix this I removed the native aliases from webgl2.c, since these
exact same aliases already exist in library_webgl.js.

This think is also more correct since it means that the `emscripten_xx`
wrappers are defined fully in JS and are not effected by definitions
in native code (which is the whole point of them I believe).