Support JSPI'd pthread entry function.
- Adds tracking of which exports are JSPI functions so we can later
determine if a function in the function table is a JSPI function.
- Adds handling for when the pthread entry function is a JSPI function and
needs a wrapper created for it.
- Handles exiting the thread when the entry function resolves.

To use JSPI + pthreads the user will need to export the thread entry function
with either `EMSCRIPTEN_KEEPALIVE` or `-sEXPORTED_FUNCTIONS=...`. The function
also needs to be marked as JSPI with `-sASYNCIFY_EXPORTS=...`