[WasmFS] Ignore rather than asserting on O_NONBLOCK
O_NONBLOCK serves two purposes when passed to open. First, it prevents `open`
from blocking for a "long time" to open the file. Second, it opens the file in
nonblocking I/O mode which allows reads and writes to return error codes if I/O
cannot be done without blocking. Ignoring O_NONBLOCK completely is a
semantically valid implementation strategy, so for now just stop asserting when
it is used.