Avoid timed waits in test_pthread_proxying_canceled_work (#19083)
In the test cases that proxy work and then exit or cancel the worker thread
before the proxied work can be executed, avoid using a timed wait to determine
when to cancel or exit the thread. Instead, signal the worker by setting an
atomic flag to tell it that the proxied work has been sent. This requires
removing the synchronous proxying from these test cases, but that's fairly safe
because the synchronous and asynchronous proxying methods have the same
internals and the cancellation is tested in combination with synchronous
proxying in other test cases.

This will fix the flake reported in #18921 if it was caused by the overloaded CI
machines taking more than 20ms to execute the proxying calls.