Postmessage worker's errors to main thread
Currently, if there is an uncaught exception in a worker thread, even if
there is an event listener for errors in the main thread, the
resulting `ErrorEvent` that reaches the main thread's event handler
doesn't have the originally thrown exception in its `error` property.

This adds another event handler in the worker side that creates an
object with the original `message` and `stack` and postmessages it to
the main thread, so the user can have more info for debugging in the
main thread.