Simplify embind $new_ helper. NFC (#19109)
This function is only needed when DYNAMIC_EXECUTION is defined.  Its
also only got two callsites that both pass `Function` as arg0.

I guess maybe this function previously had more callers that could
explain why it seems to general but I looked back as far as the big
refactor in 2014 (350f199fb9e1) and I couldn't find anything.

I think this is can be even further simplified but this is a good first
step.