Support OPFS on the main thread with JSPI. (#19063)
The OPFS backend will now use async reaad/writes when used on the main
thread. The ProxyWorker is stubbed out and calls are directly dispatched
into JS. The JS side uses a new FileSystemAsyncAccessHandle class that matches
the sync version of the class, but has async read/writes. Hopefully, we
can one day replace that with something from the platform.