Convert deps_info.py entries into JS `__deps` properties. NFC
Split out from #18905

The idea behind this change is that once we remove the `deps_info.py`
system these `__deps` entries will contain sufficient dependency
information to avoid any regressions.

Note that the symbols in `deps_info.py` don't always match the JS
library symbol names since that `deps_info.py` only contains symbols
that use code directly calls, no the lower level, internal, symbols that
the JS library code often contains.

Because of this mismatch the conversion from one form to the other is
not exact, and not one-to-one, so there could be gaps here we are
relying a lot on test coverage in #18905 to ensure we have all deps
covered.

Also note that duplicating deps between `deps_info.py` and `__deps`
entires is something we have been doing for a while now and can be
independently useful.