Deprecate the USES_DYNAMIC_ALLOC setting (#19164)
All this setting was doing was setting MALLOC=none so it doesn't seem
like it needs to exist as a separate setting.

I think it used to do more, for example, prior to #12790 it used was
used in some JS library code.  Another PR that removed a lot of usage
of this settings was #12057.

Simply removing it simplifies another change I'm working on: #18905.

Also, the docs were incorrect in stating it was ignored except under
MINIMAL_RUNTIME, but we had no testing for it.