Make sure that error() always generates failure in `compiler.js`
Previously we would report the error but then exit with 0 in some
cases.  For example in `--symbols-only` mode of if the error happened
after the first check of `abortExecution` in this file.

Also, change to use `Unknown preprocessor directive` other it will
be reported as "Internal compiler error", which it is not.