Update minimum Node.js version to 15.0.0
This is about 2 years old, and the node website already provides newer versions both
for latest and for LTS. However, other places are behind, for example the latest
Ubuntu LTS (22) has Node 12, so perhaps we should wait on this.

This is possible after https://github.com/emscripten-core/emsdk/pull/829 made the emsdk
install a 15.x version by default.