Set minimum node version in emcc invocation in tests and gen_struct_info (#19193)
This is not needed now, but when we switch to wasm-bigint by default then we'd
lose the ability to run in older nodes without it. That is, with this change, we tell
emcc the version of node we'll be running in, so it can decide properly on how to
build (and if the version is too old, we'll disable wasm-bigint).