Remove fallback to BlobBuilder if Blob constructor is missing
This code for handling the absence of Blob constructor goes back to
2012: a97aab0522858f3e24ab96fe93e7be9e2b976878

I'm pretty there is nobody depending on this today.  I think we would
have to go back to chrome 32 or firefox 14 or safari 6 in order to
actually need this: https://caniuse.com/blobbuilder

I noticed this looked like it should be removed when following the
link in tools/file_packager.py to https://crbug.com/124926, which
was marked as WontFix over 10 years ago.