Add Opera & Edge as a browsers that support JSPI
Opera and Edge are Chromium-based, so they also support JSPI (since O96).
Use the current simple detection algorithm based on executable full path
part pattern detection, but extend it with a new pattern for Opera and
Edge.

Also, print the available options for running test_async when a
JSPI-capable browser was not provided.