Stop requiring DYNAMIC_EXECUTION when building with RELOCATABLE
I tracked this error all the way back to when it was first added back
in 2015: 38157ef59e42b01078164098154baf78afbf8fc7.

Its not longer true that `loadDynamicLibrary` requires `eval()`.

It is true that if you have EM_ASM in your side modules that would
require `eval()` but that usage is guarded using the `makeEval` helper
(and also not common).