Add TLS size to allocation in emscripten_malloc_wasm_worker
This fixes `lsan.test_wasm_worker_hello`.  The reason this was failing
is that lsan adds about 26k of TLS data which meant that 1024 bytes of
total allocation was not nearly enough.

Ideally we would fix this for `emscripten_create_wasm_worker` too.  We
could either do this by saying that the TLS region is always separetly
allocated or we could provide some way to for the user to discover
the TLS size (i.e. a wrapper around __builtin_wasm_tls_size) so that
can add it to their own allocation.