Keep DSO file data around after dlsync. NFC
Its not generally safe to free this data because some threads might
have been alseep when the dlsync occurred.  In these cases they need
to be able to catchup/sync when they wake at some point in the future.

This also happens to fix the issue where new threads that get created
after the dlopen didn't have access to the file data.

The downside is that, for now, we leak the file data for DSOs that
get loaded.

Fixes: #19371