Remove WASM2C option (#19380)
This was very useful for me personally in exploring the options in the wasm2c
space, and in benchmarking in particular, for:

https://kripken.github.io/blog/wasm/2020/07/27/wasmboxc.html

At this time, however, all known users of wasm2c are using upstream wasm2c
from wabt directly. And the benefits of emscripten's integration are diminishing
as features like wasm-EH remove the need for emscripten-EH (which wasm2c
supported here). So we might as well remove this option to get rid of the
complexity, since this is already quite behind upstream wasm2c, and the work
to update it is not trival, and not worth it given the lack of users.