Ensure system libraries are built on demand
Previously, when explicitly linking against a system library (e.g.
`-lc`) that had not been built earlier, wasm-ld would raise an error
indicating that the library could not be found.

Use the `get_link_flag()` function to ensure that it triggers a
build if the system library is not present in the cache.