WasmFS: Add basic NODERAWFS support (#19400)
This PR makes the NODERAWFS flag make the root directory be
created using the Node backend. That may not be 100% of what we
need for NODERAWFS but it gets a significant amount of tests
passing.

To do that, add a hook that backends can use to override creation of
the root directory. Then NODERAWFS support basically is just to link
in a tiny library that overrides that hook.

This adds a new test, test_noderawfs_wasmfs which verifies we write
files out to the normal filesystem. Also, enable test_freetype, which
now passes; previously it failed during a configure step (specifically,
configure tried to write the size of an integer to a file... before this
PR, the file was empty and it defined integer size as the empty string).
So test_freetype both verifies NODERAWFS writes, and also it uses
files internally (to read font files etc.) which seems useful to enable.

Also, this is tested in test_fs_writeFile_rawfs. That test does
FS.writeFile with NODERAWFS enabled. It passed before, because
enabling NODERAWFS did nothing... for WasmFS the test really just
worked on the MemoryFile backend, since it doesn't create a Node
backend. With this PR, the test properly uses Node files, and passes
(thanks to #19397 and #19396).

Note: other.test_unistd_fstatfs_wasmfs was removed from CI because
it was unnecessary - all other tests run anyhow. We only need to add
select wasmfs tests because that mode does not run in full already.