WasmFS: Fix __syscall_utimensat to handle nonstandard linux behavior (#19421)
Musl and Linux allow that syscall to get a null path and to then behave
as if AT_EMPTY_PATH were passed, but without passing that flag. That is,
if the path is null then the syscall is to operate on the FD (which might be
anything, and not just a directory, as the parameter normally is).

Another option to fix this might be to modify musl's source code to not
do that in futimens or utimensat but I think it's better to avoid modifing
musl source code.

With this, wasmfs.test_futimens passes.