WasmFS: Avoid malloc in FS.readFile (#19435)
The dependency on malloc there caused a WASI test to fail. It looks like the
WASI test binary doesn't have malloc in it, but this is a post-link test so we
can't add malloc. In any case, we don't use malloc in other similar places in
the file, so switch to a stack allocation which also gets the test passing.

Also add a missing dep on malloc in FORCE_FILESYSTEM mode, as the
advanced JS APIs do actually require it.