embind: Generate TypeScript definitions from exported bindings.
Adds a new flag `--embind-emit-tsd <path>` that creates an alternative
version of the program that is run in node to collect all
the bindings and generate a TypeScript definition file.

The TS version of the program works by replacing `embind.js`
with a new file `embind_ts.js` that captures all the bindings. Much of
the type dependency code was taken from `embind.js` and moved into a new
shared file that both versions use.

Notes: This is the start of the TS gen code and I have nearly all of the
other types done, but wanted to start here for review.  In the future, I
I also want to improve it so that the user only compiles once with the
flag and the TS file is generated on the side, but the final build is a
normal build.