WasmFS: Support LSan (#19443)
This is a rather tricky problem, since

* LSan needs WasmFS to be alive in order to printf.
* WasmFS allocates during startup.
* LSan starts up at the first allocation, which ends up inside WasmFS, so LSan starts up first.
* As a result LSan shuts down last due to the order of atexits, which is after WasmFS shut down, so it can't print.

That is, the order of events is

* WasmFS begins to start up, and allocates an innocent vector or such.
* LSan gets called from the allocation and it starts up and does an atexit.
* WasmFS finishes starting up and does its own atexit for its destructor automatically.
*  WasmFS's atexit gets called.
*  LSan's atexit gets called, but at this time without WasmFS being alive we can't print.

I considered other options here, like some special destructor / atexit
reordering, but everything was much more complex. The other other viable
option I can think of is to make LSan not print using printf, e.g., if it
printed using emscripten_console_* then that might work, at least on the
web, but not in standalone mode, and also it would be an intrusive change
to the LSan codebase.