WasmFS: Do not include library_syscall.js (#19442)
99% of that file is FS-related syscalls which WasmFS does not want anyhow.

The remaining 2 syscalls, newselect and fadvise, can be implemented in a
trivial way for now. (newselect can't work anyhow until we have sockets
support in WasmFS; and fadvise was already trivial in JS.)

With that in place, a test checking MAIN_MODULE=1 + WasmFS is added
which verifies there are no linker errors due to this refactoring.

This PR makes progress towards building MAIN_MODULE=1 + WasmFS +
closure, but that does not pass yet.

This PR includes a movement of wrapSyscallFunction to a more central
location, which is now necessary as the old position in library_syscall is
no longer included in all builds.