Fix test_pthread_dlopen_many
When a thread sync was cancelled (due to a thread exit prior to
performing the task) we were treating that as an error, but we should
treat that as success.

I was able to reproduce the issue by running this test a few times in
a row.  I confirmed that it did not happen in 160 runs after this fix.

Fixes: #18887