Docs: Fix emscripten_sleep docs which referred to old 'yield' behavior (#19473)
Historically in the Emterpreter days we used to have a "yield" behavior, in which
a sleep could either yield or not. If it yielded then async events could happen; if
not then they'd be queued up for after it. That was implemented using manual
hooks in SDL and a few other places, so it wasn't a general solution, and was
complex, so we decided not to reimplement something similar for the
Emterpreter's replacement, Asyncify. But the old docs here referred to the old
behavior of that yield mode, which this PR fixes.