Always virtualize file descriptors when using NODERAWFS
At least for the `dup*()` syscalls, virtualizing the file
descriptors is necessary due to the absence of an API for these
functions in Node.js.

This commit ensures that a virtualized file descriptor is always
returned, even for the `open()` system call and similar functions.

While doing that, it revealed a few regressions in the NODERAWFS
implementation of `fchmod()`, `fchown()` and `ftruncate()` as they
were not remapping virtualized file descriptors to their native
counterparts.