Remove unneeded normalize_line_endings function. NFC
Now that we use using python3 all inputs will be converted internally to
`\n` when files are read in.  Unless we are reading text files in binary
mode the `\r\n` ending will never exist in python, and since strings
(unicode) and binary are different types now we know that we are not
dealing with binary here.