Don't abort on dlopen/dlsym when dynamic linking is not available
Instead simply return failure with a error message in `dlerror`.  This
matches the bahaviour of other libc/syscall functions and allows the
behaviour to be controlled via the existing ALLOW_UNIMPLEMENTED_SYSCALLS
setting.

The only downside (IIUC) is that emscripten is a little less
helpful/magical in advising of what to do if these functions are used
without `-sMAIN_MODULE`.

Fixes: #19411